+++ core/impl/src/main/java/com/blazebit/persistence/impl/query/ReturningModificationQuerySpecification.java
@@ -47,13 +47,13 @@ public class ReturningModificationQuerySpecification<T> extends CustomQuerySpeci
     @Override
     public ModificationQueryPlan createModificationPlan(int firstResult, int maxResults) {
         final String sql = getSql();
-        return new CustomReturningModificationQueryPlan<T>(extendedQuerySupport, serviceProvider, exampleQuery, objectBuilder, participatingQueries, sql, firstResult, maxResults);
+        return new CustomReturningModificationQueryPlan<T>(extendedQuerySupport, serviceProvider, exampleQuery, objectBuilder, participatingQueries, sql, firstResult, maxResults, returningColumns.length == 1 && objectBuilder != null);
     }
 
     @Override
     public SelectQueryPlan createSelectPlan(int firstResult, int maxResults) {
         final String sql = getSql();
-        return new CustomReturningModificationQueryPlan<T>(extendedQuerySupport, serviceProvider, exampleQuery, objectBuilder, participatingQueries, sql, firstResult, maxResults);
+        return new CustomReturningModificationQueryPlan<T>(extendedQuerySupport, serviceProvider, exampleQuery, objectBuilder, participatingQueries, sql, firstResult, maxResults, returningColumns.length == 1 && objectBuilder != null);
     }
 
     @Override